+++ src/objects/small/Command.js
@@ -1,0 +1,26 @@
+const EventEmitter = require('events')
+
+/**
+ * Command emitter for ifunny client's
+ * Works similarly to a normal EventEmitter,
+ * but registers every listener as a command for it's client
+ * @extends {EventEmitter}
+ */
+class Command extends EventEmitter {
+    constructor(client) {
+        super()
+        this.client = client
+    }
+
+    /**
+     * Register a listener, and add it's name to the list of client Commands
+     * @param  {String}     type     Command name to be searched for following a prefix
+     * @param  {Function}   listener Function to be executed when this command is called
+     */
+    on(type, listener) {
+        super.on(type, listener)
+        this.client._commands.add(type)
+    }
+}
+
+module.exports = Command