+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-### 1.0.0
+### 0.1.1
+- Add `Chat` methods `read`, `add_operator`, `remove_operator`, `join`, `exit`, `invite`, and `kick`
+- Add `Message` methods `delete` and `edit`
+- Add `Invite` methods `accept` and `decline`
+- Add `ChatUser` get method and getters
+
+### 0.1.0
 - Add basic (mostly read-only) classes for objects
 - `Client.login` method
 - `Client.notifications_paginated` and `client.notifications`